requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==25.3.0
#backports.zoneinfo==0.2.1
billiard==4.2.1
celery==5.5.3
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
django==4.2.23
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1
importlib-resources==6.4.5
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kombu==5.5.4
packaging==25.0
pillow==10.4.0
pkgutil-resolve-name==1.3.10
prompt-toolkit==3.0.51
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
redis==6.1.1
referencing==0.35.1
rpds-py==0.20.1
six==1.17.0
sqlparse==0.5.3
typing-extensions==4.13.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
zipp==3.20.2
